fixup! tryfix: when a frame's viterbi bit errors are >30, don't copy it to the audio stream

+++ openrtx/include/protocols/M17/M17FrameDecoder.hpp
@@ -145,7 +145,7 @@ private:
     ///< Maximum allowed hamming distance when determining the frame type.
     static constexpr uint8_t MAX_SYNC_HAMM_DISTANCE = 4;
 
-    static constexpr uint32_t VITERBI_BIT_ERROR_THRESHOLD = 30;
+    static constexpr uint16_t VITERBI_BIT_ERROR_THRESHOLD = 30;
 };
 
 }      // namespace M17
